requirements.txt
langchain
langchain-openai
langchain-chroma
langchain-community
langgraph
python-dotenv
